Simulated progression data

A demo dataset containing simulated data to compute progression-free survival. For demonstration purposes only.

simprog

Format

A data frame with simulated patient data for PFS analysis

Examples

data(simprog)
head(simprog)
#> # A tibble: 6 × 10
#>   patientid baseline_date last_activity_date death_date visit_gap_start_date
#>       <int> <date>        <date>             <date>     <date>              
#> 1         1 2010-01-01    2010-08-29         2010-11-19 2010-04-07          
#> 2         2 2010-01-01    2011-07-14         NA         NA                  
#> 3         3 2010-01-01    2012-04-29         2012-05-15 NA                  
#> 4         4 2010-01-01    2010-03-05         2010-03-05 NA                  
#> 5         5 2010-01-01    2010-08-18         2010-09-11 NA                  
#> 6         6 2010-01-01    2010-02-20         2010-03-11 NA                  
#> # ℹ 5 more variables: last_progression_abstraction_date <date>,
#> #   progression_date_all_events <date>, progression_date_discard_le_14d <date>,
#> #   progression_date_radiographic_only <date>,
#> #   progression_date_no_pseudoprogression <date>